style(LinkAddBar) : 배경 , input 태그 스타일 설정
