+++ public/style.css
@@ -12,16 +12,17 @@
   --gray70: #333;
   --gray40: #666;
 }
-* {
-  box-sizing: border-box;
-  font-family: Pretendard;
-  margin: 0;
-  padding: 0;
-}
 @font-face {
   font-family: 'Pretendard-Regular';
   src: url('https://cdn.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff')
     format('woff');
   font-weight: 400;
   font-style: normal;
+}
+
+* {
+  box-sizing: border-box;
+  font-family: Pretendard;
+  margin: 0;
+  padding: 0;
 }
