+++ src/components/GradationButton.js
@@ -1,0 +1,4 @@
+function GradationButton({ text }) {
+  return <button>{text}</button>;
+}
+export default GradationButton;
