--- src/components/LinkAddBar.js
@@ -1,4 +1,0 @@
-function LinkAddBar() {
-  return <></>;
-}
-export default LinkAddBar;
