--- src/pages/FolderPage.js
@@ -1,13 +1,0 @@
-import Footer from '../components/Footer';
-import Header from '../components/Header';
-import Profile from '../components/Profile';
-import CardList from '../components/CardList';
-function FolderPage() {
-  return (
-    <>
-      <Header></Header>
-      <Footer></Footer>
-    </>
-  );
-}
-export default FolderPage;
