+++ src/pages/FolderPage/FolderPage.js
@@ -1,0 +1,14 @@
+import Footer from '../../components/Footer';
+import Header from '../../components/Header';
+import LinkAddBar from './components/LinkAddBar';
+
+function FolderPage() {
+  return (
+    <>
+      <Header></Header>
+      <LinkAddBar></LinkAddBar>
+      <Footer></Footer>
+    </>
+  );
+}
+export default FolderPage;
