+++ src/pages/SharedPage/SharedPage.js
@@ -1,0 +1,15 @@
+import Footer from '../../components/Footer';
+import Header from '../../components/Header';
+import Profile from './components/Profile';
+import CardList from '../../components/CardList';
+function Shared() {
+  return (
+    <>
+      <Header></Header>
+      <Profile></Profile>
+      <CardList></CardList>
+      <Footer></Footer>
+    </>
+  );
+}
+export default Shared;
